float -> double in athena dtypes

+++ helper/util.py
@@ -283,7 +283,7 @@ def get_athena_dtypes(maps: Dict[str, str] = {}) -> Dict[str, str]:
         elif(dtype == 'int' or dtype == 'bigint' or dtype == 'integer' or dtype == 'smallint' or dtype == 'bigserial' or dtype == 'smallserial' or dtype == 'serial' or dtype.startswith('serial') or dtype.startswith('int')):
             athena_types[key] = 'bigint'
         elif(dtype == 'float' or dtype == 'double precision' or dtype.startswith('numeric') or dtype == 'real' or dtype == 'double' or dtype == 'money' or dtype.startswith('decimal') or dtype.startswith('float')):
-            athena_types[key] = 'float'
+            athena_types[key] = 'double'
     return athena_types
